feat: simplify libro dto contracts

+++ src/main/java/com/io/librotech/dto/LibroResponseDTO.java
@@ -1,5 +1,6 @@
 package com.io.librotech.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -9,7 +10,7 @@ public record LibroResponseDTO(
         String autor,
         String isbn,
         LocalDate fechaPublicacion,
-        Double precio,
+        BigDecimal precio,
         String nombreEditorial, // Aplanamos la relación: solo enviamos el nombre
         Set<String> generos // Solo los nombres de los géneros
 ) {}
